* Ranged attack speed mods and haste bonus based on wowwiki, http://arcemu.org/forums/index.php?showtopic=2462

+++ src/arcemu-world/Player.h
@@ -1343,7 +1343,7 @@ public:
 	void EventReduceDrunk(bool full);
 	bool m_AllowAreaTriggerPort;
 	void EventAllowTiggerPort(bool enable);
-	int32 m_rangedattackspeedmod;
+	float m_rangedattackspeedmod;
 	int32 m_meleeattackspeedmod;
 	uint32 m_modblockabsorbvalue;
 	uint32 m_modblockvaluefromspells;
